pyproject.toml
[project]
name = "content-automation-platform"
version = "1.0.0"
description = "A content automation platform"
dependencies = [
    "apscheduler>=3.11.0",
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.2",
    "httpx==0.27.0",
    "openai>=1.102.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.23.0",
    "werkzeug>=3.1.3",
    "flask-cors>=6.0.1",
    "requests>=2.32",
    "pillow>=10.3",
    "anthropic>=0.67.0",
    "sqlmodel>=0.0.25",
    "python-jose[cryptography]>=3.5.0",
    "loguru>=0.7.3",
    "python-multipart>=0.0.20",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "twilio>=9.8.1",
    "boto3>=1.40.34",
    "cryptography>=46.0.1",
    "replicate>=1.0.7",
]
